fix: Click pause button working
By default the `BigPlayButton` is only appears at first time loaded.
After the button clicked, the button will be dissappear. This button
also only for to play naturally. If we click it again, it looks not
working, but, it's actually working because it's just to play.

Adding new feature to make it able to pause the video. This
`BigPlayButton` use because the play button in the controls is not
flexible for reposition.

Additional style applied for PiP button in controls. The PiP button
is on controls bar in Chrome, but not in Firefox.